projects/plan2blend/requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pdf2image>=1.16.3
poppler-utils>=0.1.0

# Geometry processing
shapely>=2.0.0
scipy>=1.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data handling
pyyaml>=6.0.1

# Optional: For IA vectorization (can be replaced with custom solution)
# scikit-image>=0.21.0
# torch>=2.0.0  # If using ML models
